hotfix: (Temp) uint, int incompatible errors fixed

+++ contracts/interfaces/IPriceManager.sol
@@ -3,13 +3,13 @@
 pragma solidity ^0.8.0;
 
 interface IPriceManager {
-    function setPrice(uint[] calldata, int256[] calldata) external;
+    function setPrice(uint256[] calldata, uint256[] calldata) external;
 
     function getPriceBuffer(uint) external view returns (int256);
 
-    function getIndexPrice(uint) external view returns (int256);
+    function getIndexPrice(uint) external view returns (uint256);
 
-    function getMarkPrice(uint) external view returns (int256);
+    function getMarkPrice(uint) external view returns (uint256);
 
     function getAverageExecutionPrice(
         uint,
